requirements.txt
dj-database-url==0.5.0
Django==2.2.1
django-appconf==1.0.3
django-compressor==2.2
django-dropbox-storage==0.4.1
django-heroku==0.3.1
django-mathfilters==0.4.0
django-sass-processor==0.7.3
django-storages==1.7.1
gunicorn==19.9.0
libsass==0.18.0
Pillow==5.1.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
pytz==2019.1
rcssmin==1.0.6
rjsmin==1.0.12
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
urllib3==1.24.3
whitenoise==4.1.2